fix: directory 구조 변경
app route와 component 구분을 위한 변경
