+++ app/page.tsx
@@ -1,0 +1,3 @@
+export default function Home() {
+  return <main></main>
+}
